refactor: Relocate LicensesActivity and introduce LicensesScreen
- Moved `LicensesActivity` from `app.about.ui` to `app.licenses`.
- Created a new `LicensesScreen.kt` composable within the `app.licenses` package.
- `LicensesActivity` now utilizes `LicensesScreen` to display its content.
- `LicensesScreen` uses `LargeTopAppBarWithScaffold` for its layout and integrates `LibrariesContainer` from the "AboutLibraries" library to display open-source licenses.
- Updated `AndroidManifest.xml` to reflect the new path for `LicensesActivity`.
- Updated `IntentsHelper.kt` to import `LicensesActivity` from its new location.
