+++ apptoolkit/src/main/java/com/d4rk/android/libs/apptoolkit/app/licenses/LicensesActivity.kt
@@ -1,0 +1,28 @@
+package com.d4rk.android.libs.apptoolkit.app.licenses
+
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.ui.Modifier
+import com.d4rk.android.libs.apptoolkit.app.theme.style.AppTheme
+
+class LicensesActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            AppTheme {
+                Surface(
+                    modifier = Modifier.Companion.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    LicensesScreen(activity = this@LicensesActivity)
+                }
+            }
+        }
+    }
+}
